dtype

+++ ringattention/ringattention_pallas_tpu.py
@@ -586,11 +586,12 @@ def _flash_attention_kernel_single_batch(
                     raise NotImplementedError(
                         f"{head_dim=} should be a multiple of {MIN_BLOCK_SIZE} if larger"
                     )
-            l_scratch_ref[batch_idx] = l_next.astype(l_scratch_ref.dtype)
+            l_next = l_next.astype(l_scratch_ref.dtype)
+            l_scratch_ref[batch_idx] = l_next
             m_scratch_ref[batch_idx] = m_next.astype(m_scratch_ref.dtype)
 
             l_next_inv_safe = jnp.where(l_next == 0.0, 1.0, 1.0 / l_next)
-            acc_scratch_ref[batch_idx] *= l_broadcast(l_corr * l_next_inv_safe)
+            acc_scratch_ref[batch_idx] *= l_broadcast(l_corr * l_next_inv_safe).astype(acc_scratch_ref.dtype)
             v = pl.load(
                 v_tile_ref, (*batch_idx, pl.dslice(start_k, block_k), slice(None))
             )
